GHC hints

+++ src/S.hs
@@ -64,7 +64,7 @@ _ ≺ _                   = False
 ι c (B _ Lt) as            = ib c (<) as
 ι c (B _ Doll) (Q _ a:as)  = r c (aas a) as
 ι c (B _ Dip) (Q _ f:a:as) = a:r c (aas f) as
-ι _ (L _ (S p)) a          = let (1,n) = UA.bounds p; (x,a_)=splitAt (fromIntegral n) a in gp p x++a_
+ι _ (L _ (S p)) a          = let (1,n) = UA.bounds p; (x,a_)=splitAt n a in gp p x++a_
 ι _ a@L{} as               = a:as
 ι _ a@Q{} as               = a:as
 ι _ a@C{} as               = a:as
